# Fix issue where doublestar globbing only worked at a single level