--- a/operator/builtin/input/file/config_test.go
+++ b/operator/builtin/input/file/config_test.go
@@ -539,7 +539,7 @@ func TestBuild(t *testing.T) {
 			require.NoError,
 			func(t *testing.T, f *InputOperator) {
 				require.Equal(t, f.OutputOperators[0], fakeOutput)
-				require.Equal(t, f.Include, []string{"/var/log/testpath.*"})
+				require.Equal(t, f.finder.Include, []string{"/var/log/testpath.*"})
 				require.Equal(t, f.FilePathField, entry.NewNilField())
 				require.Equal(t, f.FileNameField, entry.NewAttributeField("file.name"))
 				require.Equal(t, f.PollInterval, 10*time.Millisecond)