… internal/tools/go.sum
… operator/builtin/input/file/file_test.go