See [opentelemetry-collector-contrib#4628](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/4628)

The PR also refactors file finding into a more testable structure, and adds tests.