@djaglowski 2021-09-16
The fix boiled down to switching from `filepath.Glob` to `doublestar.Glob`, within `file_input`'s `poll` loop. Because of this, performance is a concern. However, benchmarks appear to show a small improvement:

This branch:
```
> make bench
go test -benchmem -run=^$ -bench ^* ./...
...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-log-collection/operator/builtin/input/file
cpu: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
BenchmarkFileInput/Single-8               468024              2194 ns/op             640 B/op          9 allocs/op
BenchmarkFileInput/Glob-8                 118584             10311 ns/op            2570 B/op         36 allocs/op
BenchmarkFileInput/MultiGlob-8            118840             11821 ns/op            2572 B/op         36 allocs/op
BenchmarkFileInput/MaxConcurrent-8         80212             12729 ns/op            2564 B/op         36 allocs/op
BenchmarkFileInput/FngrPrntLarge-8        472976              2140 ns/op             641 B/op          9 allocs/op
BenchmarkFileInput/FngrPrntSmall-8        641103              1933 ns/op             640 B/op          9 allocs/op
```

main:
```
> make bench
go test -benchmem -run=^$ -bench ^* ./...
...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-log-collection/operator/builtin/input/file
cpu: Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz
BenchmarkFileInput/Single-8               451371              2344 ns/op             640 B/op          9 allocs/op
BenchmarkFileInput/Glob-8                 114730             12301 ns/op            2573 B/op         36 allocs/op
BenchmarkFileInput/MultiGlob-8            115032             12374 ns/op            2573 B/op         36 allocs/op
BenchmarkFileInput/MaxConcurrent-8         80317             12709 ns/op            2563 B/op         36 allocs/op
BenchmarkFileInput/FngrPrntLarge-8        629154              1916 ns/op             641 B/op          9 allocs/op
BenchmarkFileInput/FngrPrntSmall-8        656941              2148 ns/op             640 B/op          9 allocs/op
```